*** empty log message ***
git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/trunk/geant4@20273 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/visualization/OpenInventor/src/G4OpenInventorViewer.cc
@@ -207,8 +207,8 @@ void G4OpenInventorViewer::SetView () {
   fSoCamera->pointAt(sbTarget);
 
   //fSoCamera->height.setValue(10);
-  fSoCamera->nearDistance.setValue(pnear);
-  fSoCamera->farDistance.setValue(pfar);
+  fSoCamera->nearDistance.setValue((float)pnear);
+  fSoCamera->farDistance.setValue((float)pfar);
   //fSoCamera->focalDistance.setValue((float)cameraDistance);
 }
